+++ providers/openfeature-provider-flagd/src/openfeature/contrib/provider/flagd/resolvers/__init__.py
@@ -1,54 +1,5 @@
-import typing
-
-from typing_extensions import Protocol
-
-from openfeature.evaluation_context import EvaluationContext
-from openfeature.flag_evaluation import FlagResolutionDetails
-
 from .grpc import GrpcResolver
 from .in_process import InProcessResolver
-
-
-class AbstractResolver(Protocol):
-    def initialize(self, evaluation_context: EvaluationContext) -> None:
-        return
-
-    def shutdown(self) -> None: ...
-
-    def resolve_boolean_details(
-        self,
-        key: str,
-        default_value: bool,
-        evaluation_context: typing.Optional[EvaluationContext] = None,
-    ) -> FlagResolutionDetails[bool]: ...
-
-    def resolve_string_details(
-        self,
-        key: str,
-        default_value: str,
-        evaluation_context: typing.Optional[EvaluationContext] = None,
-    ) -> FlagResolutionDetails[str]: ...
-
-    def resolve_float_details(
-        self,
-        key: str,
-        default_value: float,
-        evaluation_context: typing.Optional[EvaluationContext] = None,
-    ) -> FlagResolutionDetails[float]: ...
-
-    def resolve_integer_details(
-        self,
-        key: str,
-        default_value: int,
-        evaluation_context: typing.Optional[EvaluationContext] = None,
-    ) -> FlagResolutionDetails[int]: ...
-
-    def resolve_object_details(
-        self,
-        key: str,
-        default_value: typing.Union[dict, list],
-        evaluation_context: typing.Optional[EvaluationContext] = None,
-    ) -> FlagResolutionDetails[typing.Union[dict, list]]: ...
-
+from .protocol import AbstractResolver
 
 __all__ = ["AbstractResolver", "GrpcResolver", "InProcessResolver"]
