+++ providers/openfeature-provider-flagd/src/openfeature/contrib/provider/flagd/resolvers/grpc.py
@@ -16,7 +16,7 @@ from openfeature.flag_evaluation import FlagResolutionDetails
 from ..config import Config
 from ..flag_type import FlagType
 from ..proto.schema.v1 import schema_pb2, schema_pb2_grpc
-from . import AbstractResolver
+from .protocol import AbstractResolver
 
 T = typing.TypeVar("T")
